+++ sdk/src/main/java/com/memfault/cloud/sdk/SendChunksCallback.kt
@@ -10,15 +10,6 @@ package com.memfault.cloud.sdk
  * The callback will be run on the main thread.
  */
 interface SendChunksCallback : Callback {
-
-    /**
-     * An error occurred: the network may have been unavailable or there may have been an issue with
-     * the request.
-     *
-     * @param sent the number of chunks that were successfully sent prior to the error.
-     */
-    fun onError(e: Exception, sent: Int)
-
     /**
      * The queue is now empty.
      *
@@ -27,10 +18,11 @@ interface SendChunksCallback : Callback {
     fun onQueueEmpty(sent: Int)
 
     /**
-     * The server was busy, please re-try after a minimum delay.
+     * The server was busy or an error occurred, please re-try after a minimum delay.
      *
      * @param delay the delay in seconds.
      * @param sent the number of chunks successfully sent prior to the busy response.
+     * @param exception the error that occurred.
      */
-    fun onRetryAfterDelay(delay: Long, sent: Int)
+    fun onRetryAfterDelay(delay: Long, sent: Int, exception: Exception)
 }