+++ sdk/src/test/java/com/memfault/cloud/sdk/internal/DirectExecutor.kt
@@ -1,0 +1,9 @@
+package com.memfault.cloud.sdk.internal
+
+import java.util.concurrent.Executor
+
+class DirectExecutor : Executor {
+    override fun execute(command: Runnable?) {
+        command?.run()
+    }
+}